🐛 Fix fatal error if no sqlite adapter is available
If the sqlite plugin is installed and active, but no sqlite is available to the php binary (e.g. if
the extension is not installed or not enabled) then that would cause a fatal error, because
getAdapter returns null and that case and trying to access getName on null doesn't work.

Closes #440