Related to #446 Only render output on whitelisted actions
A check on the renderer class alone turned out to not be sufficient,
because xhtml is rendered for wiki internal texts within an ID context
sometimes. Eg. for the login intro text.

This keeps the renderer blacklist and only uses the regexp based check
when all other checks have been passed.