Fix struct editor missing on DokuWiki master
In the big refactoring of html.php in Dokuwiki, the events like
HTML_EDITFORM_OUTPUT were removed and replaced with FORM_EDIT_OUTPUT.
See splitbrain/dokuwiki#3198

This PR adds support for the new event and maintains the support for the
legacy one. It is expected to work both on the currently released
version of DokuWiki and the current tip of the development branch.