Implement twig templating for page struct data table
Currently the struct data assigned to a page is shown as a simple table
at the top of the page. To allow easy customization of this view without
messing with the code, we add a twig template parser that looks for a
matching template file at lib/plugins/struct/templates/<schema>.twig.

In case no template is available the renderer falls back to the previous
default implementation.

Signed-off-by: Frieder Schrempf <dev@fris.de>