Add config option to show data on the top of the page
Currently the default behavior is to show the data below the first
heading if it is found, or else on the top of the page.
In addition to the existing option to always show the data on the
bottom, let's also add an option to show it always on the top.

As we want to stay compatible with existing setups that use the
'bottomoutput' option we add another boolean 'topoutput'. The former
takes precedence over the latter and only if both are unset, the
default behavior of showing the data below the first heading is used.

Signed-off-by: Frieder Schrempf <dev@fris.de>