Add more tests for struct filter variables
This adds some more tests for struct filter variables and the case
where the schema referenced by the variable is not the same as the
one assigned to the search.

This also uncovers a bug which is triggered by both schemas, the one
assigned to the search and the one referenced by the variable, having
a column with the same name.

Signed-off-by: Frieder Schrempf <dev@fris.de>