Fix bug in struct filter variable evaluation
When a search contains a struct variable expression for filtering and
the variable references a schema that is not assigned to the current
search but has a column with the same name like one from the assigned
schemas, the evaluation of the variable doesn't return the correct
value.

This patch fixes the bug by ensuring that findColumn() returns false
in case the requested schema wasn't found within the assigned schemas.

To make sure the behavior of other callers of findColumn() remains the
same, we introduce a parameter 'strict' to trigger this new behavior
in applyFilterVarsStruct().

There is also a unit test added in a separate commit that explicitely
checks this case.

Signed-off-by: Frieder Schrempf <dev@fris.de>