Apply aggregation type specific filters in render phase
The different aggregation types add their own filters to ensure that the
appropriate data is selected. This used to happen in handle() but this
meant that theses filters were only ever updated when the syntax of the
page they are used in is updated.

Since we refer to getID() to filter serial schemas based on the global
ID to enable their use in includes, we can not have that in the handle
stage, otherwise output is always filtered for the include ID.