Initial implementation of custom JOINs for simple types
This supports types that need no special processing of the column data
returned from the SQLite database and which are single-valued. Further
work will be needed to support multi-valued columns. I also need to
implement some more sophisticated logic for some of the more complex
data types. Basically, any type which overrides the `filter` method
will also need an override for the purposes of JOINing.