Remove != '' filter from queries
This effectively reverts dc7d8d08 (changes were moved to different files,
so it is more transparent to do this in a new commit).

Fixes issues with negative filters #358 #380 #656

TODO: Correct filtering on prefixed/postfixed fields, like the original commit was supposed to do, but without side effects that break normal non-prefixed fields.