Additional environment cleanup in test
Recently some tests kept failing on GitHub with a disk I/O errors on DB write.
This was happening in a specific test class with a loop doing quite a few DB write
operations in setUp(), causing the subsequent test to fail.
Removing the offending test just made the next one fail.

Adding additional setUpBeforeClass() in each setup gets rid of the problem.