BREAKING. Type Tag: leaving page empty now refers to current page
Instead of assuming a page named after the field, an empty "page"
configuration now always refers to the "current" page. Eg. the page on
which the tag is rendered.

A tag without a configured aggregation page will always link to the same
page. When shown in the page data, such a tag can not be easily looked
up anymore. OTOH it can be much easier be used in several aggregations.