Attempt to fix multi value handling with placeholders
This partly reverts and augments 131fd50

The validator call is needed to transform the given value into to proper
format (eg. an array for multivalues or an array of array for multi-page
values). This has to be done after applying place holders.

This might also fix #729

It might also break other usecase so it needs more real world testing.
Overall the bureaucracy handling is a mess, because bureaucracy is a
mess. Until splitbrain/dokuwiki-plugin-bureaucracy#303 has been
addressed, there is no good way to fix this.