Try to fix placeholder application for bureaucracy fields
We were applying replacements in setVal() but that is weirdly called
during parsing and then gets cached. But it's also called again when the
validation fails... It's all a mess.

This tries to work around the issue, but I am far from sure this hits
all the (edge) cases.

Really without a refactoring of bureaucracy it's nearly impossible
to interface cleanly with it.